requirements.txt
asgiref==3.7.2
black==24.1.1
click==8.1.7
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
uritemplate==4.1.1
wheel==0.42.0
